% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiblioButler-package.R
\docType{package}
\name{BiblioButler-package}
\alias{BiblioButler}
\alias{BiblioButler-package}
\title{BiblioButler: AI-Powered Bibliographic Reference Management}
\description{
BiblioButler is an R package for performing tasks related to bibliographic
references collection, selection, and analysis with the support of AI and
machine learning functionalities.
}
\section{Main functions}{

\itemize{
\item \code{\link{collect_references}}: Collect references from a specified source
\item \link{Add more main functions as they are implemented}
}
}

\author{
\strong{Maintainer}: Angelo D'Ambrosio \email{a.dambrosioMD@gmail.com} (\href{https://orcid.org/0000-0002-2045-5155}{ORCID})

}
\keyword{internal}
