% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_article_get_input}
\alias{check_article_get_input}
\title{Check input for article retrieval functions}
\usage{
check_article_get_input(doi = NULL, title = NULL, fun)
}
\arguments{
\item{doi}{A character string representing the Digital Object Identifier
(DOI) of an article.}

\item{title}{A character string representing the title of an article.}

\item{fun}{A character string representing the name of the calling function,
used for error messages.}
}
\value{
This function does not return a value. It stops execution with an
error message if the input is invalid.
}
\description{
This function validates the input for article retrieval functions, ensuring
that either a DOI or a title is provided, but not both.
}
\examples{
check_article_get_input(doi = "10.1000/xyz123", fun = "get_article")
check_article_get_input(title = "Example Article Title", fun = "get_article")

}
