% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_get}
\alias{col_get}
\title{Get a column from a data frame}
\usage{
col_get(col)
}
\arguments{
\item{col}{A single column to retrieve from a data frame.}
}
\value{
The specified column from the data frame. If the column does not
exist, returns NA.
}
\description{
This function is designed to retrieve a single column from a data frame. If
more than one column is passed, a warning is issued.
}
