% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_collector.R
\name{collect_references}
\alias{collect_references}
\title{Collect references from a given source}
\usage{
collect_references(source, query, limit = 100)
}
\arguments{
\item{source}{A character string specifying the source of references.}

\item{query}{A character string containing the search query.}

\item{limit}{An integer specifying the maximum number of references to collect.}
}
\value{
A data frame containing collected references.
}
\description{
This function collects bibliographic references from a specified source.
}
\examples{
collect_references("pubmed", "artificial intelligence", 10)
}
