% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossreference.R
\name{cr_make_api_call}
\alias{cr_make_api_call}
\title{Make an API Call to Crossref}
\usage{
cr_make_api_call(endpoint, method = "GET", query = NULL, headers = NULL)
}
\arguments{
\item{endpoint}{The full URL for the API request.}

\item{method}{HTTP method, default "GET".}

\item{query}{A named list of query parameters.}

\item{headers}{A named list of additional HTTP headers.}
}
\value{
The parsed JSON response.
}
\description{
This function wraps httr2 requests to the Crossref REST API.
}
