% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossreference.R
\name{cr_prepare_ids}
\alias{cr_prepare_ids}
\title{Prepare DOIs for Crossref API Requests}
\usage{
cr_prepare_ids(ids)
}
\arguments{
\item{ids}{A character vector of DOIs.}
}
\value{
A character vector of valid DOIs.
}
\description{
This function takes a character vector of DOIs, removes any NA or empty values,
and returns the cleaned vector.
}
