% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossreference.R
\name{cr_process_response}
\alias{cr_process_response}
\title{Process Crossref API Response into a Standard Data Frame}
\usage{
cr_process_response(resp)
}
\arguments{
\item{resp}{The JSON response from Crossref.}
}
\value{
A data frame with standardized columns.
}
\description{
This function converts the JSON response from Crossref into a standardized
data frame with columns for DOI, title, abstract, authors, year, journal,
publication type, and a list-column for additional IDs.
}
