% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enable_parallel}
\alias{enable_parallel}
\title{Enable parallel processing for bibliobutler operations}
\usage{
enable_parallel(workers = parallel::detectCores() - 1)
}
\arguments{
\item{workers}{Number of parallel workers to use.
Defaults to \code{parallel::detectCores() - 1}.}
}
\value{
Invisibly returns the number of workers created.
}
\description{
This convenience function enables parallel processing for bibliobutler
functions by setting up background workers.
}
\details{
This function sets up parallel workers that bibliobutler functions will
automatically use to speed up operations. If workers are already running,
this will stop them and create new ones.

To disable parallel processing, use \code{disable_parallel()}.
}
\examples{
\dontrun{
# Enable parallel processing with default number of workers
enable_parallel()

# Use specific number of workers
enable_parallel(workers = 4)

# Disable parallel processing
disable_parallel()
}

}
