% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_record_name}
\alias{generate_record_name}
\title{Generate a unique name for an article}
\usage{
generate_record_name(article_data)
}
\arguments{
\item{article_data}{A data frame containing article data, including the
\code{.authors}, \code{.title}, and \code{.year} columns.}
}
\value{
A character vector of unique record names.
}
\description{
This function takes an article data frame and generates a unique name for
each article based on the year, first author, and first word of the title.
}
