% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_article_id_type}
\alias{get_article_id_type}
\title{Get the type of article ID}
\usage{
get_article_id_type(ids)
}
\arguments{
\item{ids}{A character vector of article IDs.}
}
\value{
A character vector indicating the type of article ID, or NA if the
type cannot be determined.
}
\description{
This function determines the type of article ID based on the provided string.
It checks for PMIDs, PMCIDs, DOIs, OpenAlex IDs, and Semantic Scholar IDs.
}
\examples{
get_article_id_type(c("12345678", "PMC1234567890", "10.1000/xyz123", "W1234567890", "99e1a52b664ec2b24605ce48eed7f4ff26c1bde9"))

}
