% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex.R
\name{get_openalex_linked}
\alias{get_openalex_linked}
\title{Get linked articles (references, citations, or related) from OpenAlex}
\usage{
get_openalex_linked(ids, links = c("citations", "references", "related"))
}
\arguments{
\item{ids}{A character vector of identifiers (DOI, PMID, or OpenAlex IDs).}

\item{links}{A character vector specifying which types of linked articles to
retrieve. Any combination of \code{c("citations", "references",
  "related")}.}
}
\value{
A named list containing up to three data frames: \code{references},
\code{citations}, \code{related}. Each data frame has columns:
\itemize{
\item \code{source_id} - the queried ID
\item \code{linked_id} - an OpenAlex ID (or other ID if convertible)
}
}
\description{
Get linked articles (references, citations, or related) from OpenAlex
}
\examples{
\dontrun{
get_openalex_linked("10.1371/journal.pone.0266781")
get_openalex_linked(
  c("10.1371/journal.pone.0266781", "10.1371/journal.pone.0267149"),
  links = c("citations", "references")
)
}
}
