% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parallel_process}
\alias{get_parallel_process}
\title{Get the current number of parallel workers}
\usage{
get_parallel_process()
}
\value{
Integer specifying the number of parallel workers configured.
}
\description{
This function retrieves the number of parallel workers currently configured
for \code{bibliobutler} operations. If not explicitly set, defaults to all
available cores.
}
