% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticscholar.R
\name{get_semanticscholar_article}
\alias{get_semanticscholar_article}
\title{Retrieve Articles from Semantic Scholar API}
\usage{
get_semanticscholar_article(
  ids = NULL,
  query = NULL,
  filter_query = NULL,
  fields = NULL,
  per_page = 100,
  max_results = Inf
)
}
\arguments{
\item{ids}{A vector of Semantic Scholar Paper IDs, DOIs, ArXiv IDs, MAG IDs,
ACL IDs, PubMed IDs, or Corpus IDs.}

\item{query}{A string to search for in the Semantic Scholar database.}

\item{filter_query}{A string to filter the search results (used with advanced
query syntax).}

\item{fields}{A vector of field names to be returned in the API response. If
NULL, default fields are used.}

\item{per_page}{Number of results to return per page (default is 100, max is
100).}

\item{max_results}{Maximum number of total results to return (default is
Inf).}
}
\value{
A list of article data from Semantic Scholar.
}
\description{
This function fetches article data from the Semantic Scholar API using either
article IDs, a search query, or a filter query. It allows for customization
of returned fields and pagination of results. Only one of \code{ids}, \code{query}, or
\code{filter_query} should be provided.
}
\details{
This function interacts with the Semantic Scholar API to retrieve
article information. We suggest getting an API key from Semantic Scholar
and setting the 'bibliobutler_semanticscholar_key' option to not incur in
rate limits.
}
\examples{
\dontrun{
# Search for articles about "machine learning"
results <- get_semanticscholar_article(query = "machine learning", max_results = 500)

# Get specific articles by their IDs
results <- get_semanticscholar_article(ids = c("649def34f8be52c8b66281af98ae884c09aef38b"))
}

}
