% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticscholar.R
\name{get_semanticscholar_articles}
\alias{get_semanticscholar_articles}
\title{Retrieve Articles from Semantic Scholar API}
\usage{
get_semanticscholar_articles(
  ids = NULL,
  query = NULL,
  year_filter = NULL,
  filters = NULL,
  fields = NULL,
  per_page = 100,
  max_results = Inf
)
}
\arguments{
\item{ids}{A vector of Semantic Scholar Paper IDs, DOIs, ArXiv IDs, MAG IDs,
ACL IDs, PubMed IDs, or Corpus IDs.}

\item{query}{A string to search for in the Semantic Scholar database.}

\item{year_filter}{Publication year or range (e.g., "2019", "2016-2020",
"2010-", "-2015"). If provided, this overrides any year filter in the
filters argument.}

\item{filters}{A list of filters to apply to the search. Available filters
include:
\itemize{
\item publicationTypes (string): Filter by publication type, as a
comma-separated list,
\item openAccessPdf (string): Restricts to papers with public PDF. No
value needed.
\item minCitationCount (string): Minimum number of citations
\item publicationDateOrYear (string): Date range in YYYY-MM-DD:YYYY-MM-DD
format. Partial dates (YYYY-MM or YYYY) and open ranges supported.
\item year (string): Publication year or range (see year_filter
parameter)
\item venue (string): Filter by venue/journal names, comma-separated
list. Accepts full names or ISO4 abbreviations.
\item fieldsOfStudy (string): Comma-separated list.
}}

\item{fields}{A vector of field names to be returned in the API response. If
NULL, default fields are used. Note that "citations" and "references" are
not supported in the bulk search and will be ignored.}

\item{per_page}{Number of results per page (default is 100).}

\item{max_results}{Maximum number of total results to return (default is
Inf).}
}
\value{
A data frame of article data from Semantic Scholar.
}
\description{
This function fetches article data from the Semantic Scholar API
(https://api.semanticscholar.org/api-docs) using either article IDs (via the
\verb{/paper/batch} endpoint), a search query (via the \verb{/paper/search/bulk}
endpoint), or a filter query. It allows for customization of returned fields
and pagination of results. Only one of \code{ids} or \code{query} should be provided.
}
\details{
This function interacts with the Semantic Scholar API to retrieve
article information. We suggest getting an API key from Semantic Scholar
and setting the 'bibliobutler.semanticscholar_key' option to not incur in
rate limits.
}
\examples{
\dontrun{
# Search for articles about "machine learning"
results <- get_semanticscholar_articles(
  query = "machine learning",
  year_filter = "2023-",
  max_results = 500
)

# Get specific articles by their IDs
results <- get_semanticscholar_articles(
  ids = c(
    "649def34f8be52c8b66281af98ae884c09aef38b",
    "10.1145/3197026.3197040"
  )
)
}

}
