% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{msg_error}
\alias{msg_error}
\title{Display an error message in red}
\usage{
msg_error(..., appendLF = TRUE)
}
\arguments{
\item{...}{One or more strings to be concatenated into a message}

\item{appendLF}{Logical. Should a line feed be appended? Default TRUE}
}
\description{
Display an error message in red
}
\keyword{internal}
