% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{msg_info}
\alias{msg_info}
\title{Display an info message in blue}
\usage{
msg_info(..., appendLF = TRUE)
}
\arguments{
\item{...}{One or more strings to be concatenated into a message}

\item{appendLF}{Logical. Should a line feed be appended? Default TRUE}
}
\description{
Display an info message in blue
}
\keyword{internal}
