% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{msg_status}
\alias{msg_status}
\title{Display a process/status message in silver}
\usage{
msg_status(..., appendLF = TRUE)
}
\arguments{
\item{...}{One or more strings to be concatenated into a message}

\item{appendLF}{Logical. Should a line feed be appended? Default TRUE}
}
\description{
Display a process/status message in silver
}
\keyword{internal}
