% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{msg_success}
\alias{msg_success}
\title{Display a success message in green}
\usage{
msg_success(..., appendLF = TRUE)
}
\arguments{
\item{...}{One or more strings to be concatenated into a message}

\item{appendLF}{Logical. Should a line feed be appended? Default TRUE}
}
\description{
Display a success message in green
}
\keyword{internal}
