% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{msg_warn}
\alias{msg_warn}
\title{Display a warning message in yellow}
\usage{
msg_warn(..., appendLF = TRUE)
}
\arguments{
\item{...}{One or more strings to be concatenated into a message}

\item{appendLF}{Logical. Should a line feed be appended? Default TRUE}
}
\description{
Display a warning message in yellow
}
\keyword{internal}
