% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex.R
\name{oa_find_id}
\alias{oa_find_id}
\title{Attempt to find an OpenAlex ID URL for a given ID (which might be a DOI, PMID, etc.)
Return \code{NA} if we can't resolve it.}
\usage{
oa_find_id(x)
}
\description{
Attempt to find an OpenAlex ID URL for a given ID (which might be a DOI, PMID, etc.)
Return \code{NA} if we can't resolve it.
}
