% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex.R
\name{oa_process_response}
\alias{oa_process_response}
\title{Process the openalexR result to a standard format}
\usage{
oa_process_response(data)
}
\arguments{
\item{data}{The tibble/data.frame returned from
\code{openalexR::oa_fetch(...)} with a set of sanitized columns at the
beginning, identified by a dot prefix.}
}
\value{
A data frame with standardized columns.
}
\description{
Similar to \code{s2_process_response()}, we want columns like: paperId,
title, abstract, authors, year, journal, pubtype, .references (list),
.citations (empty by default), .related (list), .api = "openalex", .ids
(data.frame with openalex, doi, pmid, etc.)
}
