% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_authors}
\alias{parse_authors}
\title{Parse author names}
\usage{
parse_authors(authors, to_string = FALSE)
}
\arguments{
\item{authors}{A vector of author names.}
}
\value{
A data frame with parsed author names.
}
\description{
This function takes a vector of author names and parses them into a
standardized format: full family name followed by initials.
}
