% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{pm_make_request}
\alias{pm_make_request}
\title{Create a PubMed API request}
\usage{
pm_make_request(endpoint, params, method = "GET")
}
\arguments{
\item{endpoint}{The API endpoint URL.}

\item{params}{List of parameters for the request.}

\item{method}{HTTP method to use ("GET" or "POST").}
}
\value{
An httr2 response object.
}
\description{
Internal function to configure and execute API requests to PubMed.
}
\keyword{internal}
