% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticscholar.R
\name{prepare_semanticscholar_ids}
\alias{prepare_semanticscholar_ids}
\title{Prepare IDs for Semantic Scholar API}
\usage{
prepare_semanticscholar_ids(ids)
}
\arguments{
\item{ids}{A character vector of article identifiers. Can be DOIs, PMIDs, or
Semantic Scholar IDs.}
}
\value{
A character vector of properly formatted IDs for use with the
Semantic Scholar API. DOIs will be prefixed with "DOI:", PMIDs with
"PMID:", and Semantic Scholar IDs will be left as-is. Invalid or
unconvertible IDs are removed.
}
\description{
This function takes a vector of article identifiers and prepares them for use
with the Semantic Scholar API by converting them to the appropriate format.
}
\examples{
\dontrun{
# Prepare mixed IDs
ids <- c(
  "10.1016/j.cell.2023.01.001", "12345678",
  "649def34f8be52c8b66281af98ae884c09aef38b"
)
prepared_ids <- prepare_semanticscholar_ids(ids)
}

}
