% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove common English stopwords from a character vector}
\usage{
remove_stopwords(x)
}
\arguments{
\item{x}{A character vector to remove stopwords from.}
}
\value{
The input character vector with stopwords removed.
}
\description{
This function takes a character vector and removes common English stopwords
from it. The stopwords are defined in a hard-coded list within the function.
Stopword list and code adapted from the \code{tm} package
https://cran.r-project.org/package=tm
}
