% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_url_from_id}
\alias{remove_url_from_id}
\title{Remove URL prefixes from a vector of IDs}
\usage{
remove_url_from_id(ids)
}
\arguments{
\item{ids}{A vector of IDs.}
}
\value{
The input vector with URL prefixes removed.
}
\description{
This function takes a vector of IDs and removes the URL prefixes from them.
}
