% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticscholar.R
\name{s2_buld_id}
\alias{s2_buld_id}
\title{Build a Semantic Scholar ID from paper data}
\usage{
s2_buld_id(paper_data)
}
\arguments{
\item{paper_data}{A list of paper data, typically from the Semantic Scholar
API.}
}
\value{
A character vector of Semantic Scholar IDs.
}
\description{
This function takes a list of paper data and attempts to build a unique
identifier for the paper. It checks the \code{externalIds} field for DOI, ArXiv,
and PubMed IDs, and if none of those are available, it uses the \code{paperId}
field.
}
