% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parallel_process}
\alias{set_parallel_process}
\title{Set the number of parallel workers for \code{bibliobutler} operations}
\usage{
set_parallel_process(n_cores)
}
\arguments{
\item{n_cores}{Integer specifying the number of cores to use.}
}
\value{
No return value, called for side effects.
}
\description{
This function sets the number of parallel workers that \code{bibliobutler} will
use for parallel processing operations. The setting is stored in the
'bibliobutler.workers' option.
}
