% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_get}
\alias{col_get}
\title{Get a column or nested element from a data frame}
\usage{
col_get(...)
}
\arguments{
\item{...}{One or more column names or indices to access nested elements.
For example, col_get("authors", "name") would retrieve the "name" field
from the "authors" column.}
}
\value{
The specified element from the data frame. If the element does not
exist, returns NA.
}
\description{
This function retrieves a column or nested element from a data frame using
multiple column names. It uses purrr::pluck() to access nested elements and
returns NA if the element does not exist.
}
