% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticscholar.R
\name{s2_process_response}
\alias{s2_process_response}
\title{Process Response from Semantic Scholar API}
\usage{
s2_process_response(data)
}
\arguments{
\item{data}{A list or data frame containing the raw API response data from
Semantic Scholar. Can be either the full response object or just the data
portion.}
}
\value{
A data frame containing the processed paper metadata with a set of
sanitized columns at the beginning, identified by a dot prefix.
}
\description{
This function processes the raw response data from the Semantic Scholar API
and transforms it into a standardized format. It extracts and formats key
paper metadata including IDs, title, abstract, authors, year, journal,
publication type, references, and citations.
}
